+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Advancements are back
 ### Fixed
 - Soul urns should be fully functional (but cannot be crafted in survival due to lack of soul gem drops)
+- Ender crystal block levitation effect
 
 ## [3.0.3] - 2019-02-24
 ### Added